# Add array index number to output when casting from json to array.