--- a/edb/ir/ast.py
+++ b/edb/ir/ast.py
@@ -1074,7 +1074,7 @@ class TypeCast(ImmutableExpr):
     sql_function: typing.Optional[str] = None
     sql_cast: bool
     sql_expr: bool
-    error_message_context: typing.Optional[str] = None
+    error_message_context: typing.Optional[Set] = None
 
     @property
     def typeref(self) -> TypeRef: