Given the query `SELECT <array<int64>>to_json('[1, 2, "asdf"]')`, produces the error:
```
InvalidValueError: while casting 'std::json' to 'array<std::int64>', in array elements, at index 2, expected JSON number or null; got JSON string
```