@msullivan 2024-05-30
> Adding `error_message_context` in `fini_expression` is super dubious imo

Why is that necessary?