@dnwpark 2024-05-30
In `_compile_call_args` there is a check to ensure that all args have a known multiplicity. The multiplicity is added in `fini_expression`. I suppose the other alternative is to add an override to `infer_multiplicity` on `TypeCast`.