@msullivan 2024-05-30
Yeah, adding it to multiplicity and cardinality checks is the correct thing

On Thu, May 30, 2024, 07:21 dnwpark ***@***.***> wrote:

> In _compile_call_args there is a check to ensure that all args have a
> known multiplicity. The multiplicity is added in fini_expression. I
> suppose the other alternative is to add an override to infer_multiplicity
> on TypeCast.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/edgedb/edgedb/pull/7397#issuecomment-2139669092>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACTC7NSVWCINFOACRD3RTDZE4YUZAVCNFSM6AAAAABIIIVZQ6VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCMZZGY3DSMBZGI>
> .
> You are receiving this because your review was requested.Message ID:
> ***@***.***>
>