@dnwpark 2024-06-04
Benchmarking against three queries:

**Simple null** - 10000 times
```
client.query(
    "SELECT <array<int64>>to_json(<str>$input)",
    input='[null]'
)
```
- `master`: `8.927s`
- With array indexes: `13.453s`

**Long array** - 1000 times
```
client.query(
    "SELECT <array<int64>>to_json(<str>$input)",
    input='['+','.join(str(i) for i in range(1000))+',null'+']'
)
```
- `master`: `4.647s`
- With array indexes: `10.134s`

**Long and deep array** - 1000 times
```
client.query(
    "SELECT <array<tuple<array<tuple<array<int64>>>>>>to_json(<str>$input)",
    input='['+','.join('[[[[1,2,3],[4,5,6],[7,8,9]]]]' for i in range(1000))+',[[[[1,2,null]]]]'+']'
)
```
- `master`: `9.406s`
- With array indexes: `90.841s`