@dnwpark 2024-06-04
@msullivan non-trivial cases seem convincingly much slower.