@msullivan 2024-06-05
Great work experimenting with this, though!