@msullivan 2024-06-05
Also, we support `json` arguments directly too